# src/main/resources/static/api.yml
openapi: 3.1.0

info:
  title: Leaderboard API
  description: API for managing leaderboard entries
  version: 1.0.0

servers:
  - url: https://localhost:8080

paths:
  /v1/leaderboard:
    get:
      operationId: getLeaderboard
      summary: Get the leaderboard entries
      tags:
        - Leaderboard
      responses:
        200:
          description: List of leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
  /v1/leaderboard/entry:
    post:
      operationId: postLeaderboardEntry
      summary: Add a new leaderboard entry
      tags:
        - Leaderboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaderboardEntry'
      responses:
        201:
          description: Leaderboard entry created
        400:
          description: Invalid data
    put:
      operationId: putLeaderboardEntry
      summary: Update an existing leaderboard entry
      tags:
        - Leaderboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaderboardEntry'
      responses:
        204:
          description: Leaderboard entry updated successfully
        400:
          description: Invalid leaderboard entry data
        404:
          description: Leaderboard entry not found
  /v1/leaderboard/entry/{playerName}:
    delete:
      operationId: deleteLeaderboardEntry
      summary: Delete a leaderboard entry
      tags:
        - Leaderboard
      parameters:
        - name: playerName
          in: path
          required: true
          schema:
            type: string
          description: The playerName of the leaderboard entry to delete.
        - name: X-USERNAME
          in: header
          required: true
          schema:
            type: string
          description: >
            Username making the request. Must match the `playerName` being deleted or be `admin`.
      responses:
        204:
          description: Leaderboard entry deleted successfully
        401:
          description: Unauthorized - Blank X-USERNAME
        403:
          description: Forbidden - X-USERNAME does not match the playerName or `admin`
components:
  schemas:
    LeaderboardEntry:
      type: object
      properties:
        playerName:
          type: string
        score:
          type: integer